# fix(VSelect): improve selection performance