--- a/packages/vuetify/src/composables/list-items.ts
+++ b/packages/vuetify/src/composables/list-items.ts
@@ -1,11 +1,11 @@
 // Utilities
-import { computed } from 'vue'
-import { deepEqual, getPropertyFromItem, omit, propsFactory } from '@/util'
+import { computed, shallowRef, toRaw, watchEffect } from 'vue'
+import { deepEqual, getPropertyFromItem, isPrimitive, omit, propsFactory } from '@/util'
 
 // Types
 import type { PropType } from 'vue'
 import type { InternalItem } from '@/composables/filter'
-import type { SelectItemKey } from '@/util'
+import type { Primitive, SelectItemKey } from '@/util'
 
 export interface ListItem<T = any> extends InternalItem<T> {
   title: string
@@ -24,7 +24,7 @@ export interface ItemProps {
   itemChildren: SelectItemKey
   itemProps: SelectItemKey
   returnObject: boolean
-  valueComparator: typeof deepEqual
+  valueComparator: typeof deepEqual | undefined
 }
 
 // Composables
@@ -50,10 +50,7 @@ export const makeItemsProps = propsFactory({
     default: 'props',
   },
   returnObject: Boolean,
-  valueComparator: {
-    type: Function as PropType<typeof deepEqual>,
-    default: deepEqual,
-  },
+  valueComparator: Function as PropType<typeof deepEqual>,
 }, 'list-items')
 
 export function transformItem (props: Omit<ItemProps, 'items'>, item: any): ListItem {
@@ -84,10 +81,11 @@ export function transformItem (props: Omit<ItemProps, 'items'>, item: any): List
 }
 
 export function transformItems (props: Omit<ItemProps, 'items'>, items: ItemProps['items']) {
-  const array: ListItem[] = []
+  const _props = toRaw(props)
 
+  const array: ListItem[] = []
   for (const item of items) {
-    array.push(transformItem(props, item))
+    array.push(transformItem(_props, item))
   }
 
   return array
@@ -97,15 +95,44 @@ export function useItems (props: ItemProps) {
   const items = computed(() => transformItems(props, props.items))
   const hasNullItem = computed(() => items.value.some(item => item.value === null))
 
-  function transformIn (value: any[]): ListItem[] {
-    // Cache unrefed values outside the loop
+  const itemsMap = shallowRef<Map<Primitive, ListItem[]>>(new Map())
+  const keylessItems = shallowRef<ListItem[]>([])
+  watchEffect(() => {
     const _items = items.value
+    const map = new Map()
+    const keyless = []
+    for (let i = 0; i < _items.length; i++) {
+      const item = _items[i]
+      if (isPrimitive(item.value) || item.value === null) {
+        let values = map.get(item.value)
+        if (!values) {
+          values = []
+          map.set(item.value, values)
+        }
+        values.push(item)
+      } else {
+        keyless.push(item)
+      }
+    }
+    itemsMap.value = map
+    keylessItems.value = keyless
+  })
+
+  function transformIn (value: any[]): ListItem[] {
+    // Cache unrefed values outside the loop,
+    // proxy getters can be slow when you call them a billion times
+    const _value = toRaw(value)
+    const _items = itemsMap.value
+    const _allItems = items.value
+    const _keylessItems = keylessItems.value
     const _hasNullItem = hasNullItem.value
     const _returnObject = props.returnObject
-    const valueComparator = props.valueComparator
+    const hasValueComparator = !!props.valueComparator
+    const valueComparator = props.valueComparator || deepEqual
+    const _props = toRaw(props)
 
     const returnValue: ListItem[] = []
-    for (const v of value) {
+    main: for (const v of _value) {
       // When the model value is null, return an InternalItem
       // based on null only if null is one of the items
       if (!_hasNullItem && v === null) continue
@@ -113,13 +140,31 @@ export function useItems (props: ItemProps) {
       // String model value means value is a custom input value from combobox
       // Don't look up existing items if the model value is a string
       if (_returnObject && typeof v === 'string') {
-        returnValue.push(transformItem(props, v))
+        returnValue.push(transformItem(_props, v))
         continue
       }
 
-      returnValue.push(
-        _items.find(item => valueComparator(v, item.value)) || transformItem(props, v)
-      )
+      // Fast path, items with primitive values and no
+      // custom valueComparator can use a constant-time
+      // map lookup instead of searching the items array
+      const fastItems = _items.get(v)
+
+      // Slow path, always use valueComparator.
+      // This is O(n^2) so we really don't want to
+      // do it for more than a couple hundred items.
+      if (hasValueComparator || !fastItems) {
+        for (const item of (hasValueComparator ? _allItems : _keylessItems)) {
+          if (valueComparator(v, item.value)) {
+            returnValue.push(item)
+            continue main
+          }
+        }
+        // Not an existing item, construct it from the model (#4000)
+        returnValue.push(transformItem(_props, v))
+        continue
+      }
+
+      returnValue.push(...fastItems)
     }
 
     return returnValue