<!--
MAKE SURE TO READ THE CONTRIBUTING GUIDE BEFORE CREATING A PR
https://vuetifyjs.com/getting-started/contributing

Provide a general summary of your changes in the title above
Keep the title short and descriptive, as it will be used as a commit message
PR titles should follow conventional-changelog-angular:
https://vuetifyjs.com/getting-started/contributing/#commit-guidelines
-->

## Description
<!--
Describe your changes in detail. Why is this change required? What problem does it solve?
If it fixes an open issue, please link to the issue here.
e.g. resolves #4213 or fixes #2312
-->

fixes #20703

Time to run `transformIn`, primitive values:
| Item # | Before | ~~After~~ | After |
|--------|--------|--------|--------|
| 1k | 320ms | ~~56ms~~ | 0.5ms |
| 10k | 40sec | ~~780ms~~ | 1ms |
| 100k | &infin; | ~~75sec~~ | 8ms |
| 1M | &infin; | ~~&infin;~~ | 125ms |

Object values (`{ key: i }`):
| Item # | Before | After |
|--------|--------|--------|
| 1k | 900ms | 33ms |
| 10k | 88sec | 2.5sec |
| 100k | &infin; | &infin; |
| 1M | &infin; | &infin; |

return-object with item-value set appropriately uses primitive value lookup

## Markup:
<!--
Information on how to set up your local development environment can be found here:
https://vuetifyjs.com/getting-started/contributing/#setting-up-your-environment
Remove this section for documentation or test-only changes.
-->

<!-- Paste your FULL packages/vuetify/dev/Playground.vue here --->
```vue
<template>
  <v-app>
    <v-container>
      <v-autocomplete
        v-model="selectedItems"
        :items="items"
        label="test"
        multiple
      >
        <template #prepend-item>
          <v-list-item @click="selectAll()">
            <v-list-item-title> Select all </v-list-item-title>
          </v-list-item>
        </template>
        <template #selection="{ item, index }">
          <template v-if="selectedItems.length > 5">
            <span v-if="!index">{{ selectedItems.length }}</span>
          </template>
          <template v-else>
            <span v-if="index" class="hint mr-1"> or </span>
            <v-chip
              :style="{ height: '24px' }"
              size="small"
              variant="flat"
              label
            >
              {{ item.title }}
            </v-chip>
          </template>
        </template>
      </v-autocomplete>
    </v-container>
  </v-app>
</template>

<script setup>
  import { ref } from 'vue'

  const selectedItems = ref([])
  const items = Array.from({ length: 10000 }, (_, i) => ({ title: `Item ${i}`, value: i }))

  function selectAll () {
    selectedItems.value =
      items.length === selectedItems.value.length ? [] : items.map(c => c.value)
  }
</script>

```