@johnleider 2025-01-29
- Is there a reason for valueComparator vs _valueComparator?
- For selectStrategies, how did you test the differences?
- It seems that most of the performance boost comes from caching the valueComparator, the other changes don't seem to do much