@KaelWD 2025-01-30
- There was no valueComparator variable in scope already
- There's no behaviour difference, the map is created in strategy.in in the first place so cloning it for every iteration is just a waste of performance
- Caching hasNullItem and returnObject is worth 100ms with a million items. `new Map(map)` takes 24ms with just 100 items. The double filter is pretty quick but might as well save the gc on an array that's going to be thrown away instantly.