@KaelWD 2025-01-30
There's still a O(n^2) deepEqual to find the first selected item index: https://github.com/vuetifyjs/vuetify/blob/c4ba8bf3e3a4d3a6ecb6fa00b8e725b3ff16c313/packages/vuetify/src/components/VSelect/VSelect.tsx#L317-L319

And a linear search to figure out if we're adding or removing a selection: https://github.com/vuetifyjs/vuetify/blob/c4ba8bf3e3a4d3a6ecb6fa00b8e725b3ff16c313/packages/vuetify/src/components/VSelect/VSelect.tsx#L266